Move COLORS to Color.js
