+++ web/src/config.js
@@ -1,3 +1,1 @@
 export const LICHESS_API_URL = 'https://lichess.org/api/'
-
-export const deviceName = 'Launchpad X MIDI 2'
