+++ web/src/index.html
@@ -8,6 +8,7 @@
     />
     <title>LaunchChess</title>
     <link rel="icon" href="/static/logo.svg" />
+    <script src="https://cdn.jsdelivr.net/npm/webmidi@next/dist/iife/webmidi.iife.js"></script>
   </head>
   <body></body>
   <script type="module" src="index.js"></script>
